Reorganizing central config files
