+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/features/centralconfig/fetcher/ConfigurationFileProvider.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.internal.features.centralconfig;
+package co.elastic.apm.android.sdk.internal.features.centralconfig.fetcher;
 
 import java.io.File;
